Index attachments even on deleted tickets
Skipping indexing on tickets which are currently deleted may lead to
content which is later not indexed and not findable.  Even if a ticket
becomes un-deleted at a later point, it will never again be indexed, as
last_indexed() will limit to new attachments since the last run.  The
overhead of joining through Attachments and into Tickets is also not to
be overlooked -- on MySQL, it causes pessimal performance for large
systems, involving more than one filesort and temporary table.

Index all attachments, no matter the source, on the premise that
limiting to non-deleted tickets is better done at query time, not at
index time.  It is both more timely, and better limited.