Refactor PostgreSQL's insert to also do bulk insertion
PostgreSQL's full-text search can use a separate column, or a separate
table.  In the case of a separate table, the same bulk-insertion trick
as MySQL uses can also be used to noticeably speed up indexing time.

Refactor the MySQL loop to be used in both cases.  Note that instead of
attempting to parse the error message (which only works in English
locales for MySQL, and does not work at all in PostgreSQL) the rows are
instead attempted one-at-a-time.  This also better catches the case
where multiple errors occurred in one bulk insert.

This provides notable speed benefits: 5x for GIN, and 30x for GiST. Note
that this does not speed up the default configuration for PostgreSQL,
wherein the index lies in the Attachments table, and thus UPDATE
statements must be used.