Insert data to index before creating the index
Having to update the index after every insertion is a notable
performance hit; it is much more performant to create the index after
all data has been inserted.

Move the initial run of rt-fulltext-indexer to after the requisite
tables have been set up, but prior to creating the index and displaying
the required RT configuration.  This requires serializing the intended
%FullTextSearch configuration between rt-setup-fulltext-index and
rt-fulltext-indexer, as the RT_SiteConfig file will not have been
updated until after both have run; thus, serialize the configuration via
an environment variable.

The documentation is updated to note that running
rt-setup-fulltext-index is now a time-intensive operation.